chg: :truck: move _consts to consts
